test(memory-wiki): satisfy lint CI guards
